# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dtos.CityDTO:
    properties:
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      state:
        $ref: '#/definitions/dtos.StateDTO'
      state_id:
        type: integer
    type: object
  dtos.CountryDTO:
    properties:
      id:
        type: integer
      is_active:
        type: boolean
      iso2:
        type: string
      iso3:
        type: string
      name:
        type: string
      numeric_code:
        type: string
      phone_code:
        type: string
      states:
        items:
          $ref: '#/definitions/dtos.StateDTO'
        type: array
      subregion:
        $ref: '#/definitions/dtos.SubregionDTO'
      subregion_id:
        type: integer
      timezones:
        type: string
    type: object
...
  dtos.CreateCityDTO:
    properties:
      name:
        maxLength: 255
        minLength: 2
        type: string
      state_id:
        type: integer
    required:
    - name
    - state_id
    type: object
...
  dtos.CreateCountryDTO:
    properties:
      iso2:
        type: string
      iso3:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      numeric_code:
        type: string
      phone_code:
        type: string
      subregion_id:
        type: integer
      timezones:
        type: string
    required:
    - iso2
    - iso3
    - name
    - numeric_code
    - phone_code
    type: object
...
  dtos.CreateRegionDTO:
    properties:
      name:
        maxLength: 255
        minLength: 2
        type: string
    required:
    - name
    type: object
  dtos.CreateStateDTO:
    properties:
      country_code:
        type: string
      country_id:
        type: integer
      name:
        maxLength: 255
        minLength: 2
        type: string
    required:
    - country_code
    - country_id
    - name
    type: object
  dtos.CreateSubregionDTO:
    properties:
      name:
        maxLength: 255
        minLength: 2
        type: string
      region_id:
        type: integer
    required:
    - name
    - region_id
    type: object
...
  dtos.LocationSearchDTO:
    properties:
      cities:
        items:
          $ref: '#/definitions/dtos.CityDTO'
        type: array
      countries:
        items:
          $ref: '#/definitions/dtos.CountryDTO'
        type: array
      states:
        items:
          $ref: '#/definitions/dtos.StateDTO'
        type: array
    type: object
...
  dtos.RegionDTO:
    properties:
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      subregions:
        items:
          $ref: '#/definitions/dtos.SubregionDTO'
        type: array
    type: object
...
  dtos.StateDTO:
    properties:
      cities:
        items:
          $ref: '#/definitions/dtos.CityDTO'
        type: array
      country:
        $ref: '#/definitions/dtos.CountryDTO'
      country_code:
        type: string
      country_id:
        type: integer
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
    type: object
  dtos.SubregionDTO:
    properties:
      countries:
        items:
          $ref: '#/definitions/dtos.CountryDTO'
        type: array
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      region:
        $ref: '#/definitions/dtos.RegionDTO'
      region_id:
        type: integer
    type: object
...
  dtos.UpdateCityDTO:
    properties:
      is_active:
        type: boolean
      name:
        maxLength: 255
        minLength: 2
        type: string
      state_id:
        type: integer
    type: object
  dtos.UpdateCountryDTO:
    properties:
      is_active:
        type: boolean
      iso2:
        type: string
      iso3:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      numeric_code:
        type: string
      phone_code:
        type: string
      subregion_id:
        type: integer
      timezones:
        type: string
    type: object
...
  dtos.UpdateRegionDTO:
    properties:
      is_active:
        type: boolean
      name:
        maxLength: 255
        minLength: 2
        type: string
    type: object
  dtos.UpdateStateDTO:
    properties:
      country_code:
        type: string
      country_id:
        type: integer
      is_active:
        type: boolean
      name:
        maxLength: 255
        minLength: 2
        type: string
    type: object
  dtos.UpdateSubregionDTO:
    properties:
      is_active:
        type: boolean
      name:
        maxLength: 255
        minLength: 2
        type: string
      region_id:
        type: integer
    type: object
...
host: localhost:8001
info:
  contact:
    email: support@dvra.com
    name: API Support
  description: API para sistema de reclutamiento y gestión de candidatos
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: DVRA API
  version: 1.2.0
paths:
...
  /locations/cities:
    get:
      parameters:
      - description: Filter by state ID
        in: query
        name: state_id
        type: integer
      - description: Search by name
        in: query
        name: search
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.CityDTO'
            type: array
      summary: Get all cities
      tags:
      - Locations
    post:
      parameters:
      - description: City data
        in: body
        name: city
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateCityDTO'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.CityDTO'
      summary: Create city
      tags:
      - Locations
  /locations/cities/{id}:
    delete:
      parameters:
      - description: City ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: Delete city
      tags:
      - Locations
    get:
      parameters:
      - description: City ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.CityDTO'
      summary: Get city by ID
      tags:
      - Locations
    put:
      parameters:
      - description: City ID
        in: path
        name: id
        required: true
        type: integer
      - description: City data
        in: body
        name: city
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateCityDTO'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.CityDTO'
      summary: Update city
      tags:
      - Locations
  /locations/countries:
    get:
      parameters:
      - description: Filter by subregion ID
        in: query
        name: subregion_id
        type: integer
      - description: Search by name or ISO code
        in: query
        name: search
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.CountryDTO'
            type: array
      summary: Get all countries
      tags:
      - Locations
    post:
      parameters:
      - description: Country data
        in: body
        name: country
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateCountryDTO'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.CountryDTO'
      summary: Create country
      tags:
      - Locations
  /locations/countries/{id}:
    delete:
      parameters:
      - description: Country ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: Delete country
      tags:
      - Locations
    get:
      parameters:
      - description: Country ID
        in: path
        name: id
        required: true
        type: integer
      - description: Include states
        in: query
        name: include_states
        type: boolean
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.CountryDTO'
      summary: Get country by ID
      tags:
      - Locations
    put:
      parameters:
      - description: Country ID
        in: path
        name: id
        required: true
        type: integer
      - description: Country data
        in: body
        name: country
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateCountryDTO'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.CountryDTO'
      summary: Update country
      tags:
      - Locations
  /locations/countries/iso/{iso}:
    get:
      parameters:
      - description: ISO2 or ISO3 code
        in: path
        name: iso
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.CountryDTO'
      summary: Get country by ISO code
      tags:
      - Locations
  /locations/hierarchy/{id}:
    get:
      parameters:
      - description: Country ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.CountryDTO'
      summary: Get complete location hierarchy for a country
      tags:
      - Locations
  /locations/regions:
    get:
      parameters:
      - description: Include subregions
        in: query
        name: include_subregions
        type: boolean
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.RegionDTO'
            type: array
      summary: Get all regions
      tags:
      - Locations
    post:
      parameters:
      - description: Region data
        in: body
        name: region
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateRegionDTO'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.RegionDTO'
      summary: Create region
      tags:
      - Locations
  /locations/regions/{id}:
    delete:
      parameters:
      - description: Region ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: Delete region
      tags:
      - Locations
    get:
      parameters:
      - description: Region ID
        in: path
        name: id
        required: true
        type: integer
      - description: Include subregions
        in: query
        name: include_subregions
        type: boolean
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.RegionDTO'
      summary: Get region by ID
      tags:
      - Locations
    put:
      parameters:
      - description: Region ID
        in: path
        name: id
        required: true
        type: integer
      - description: Region data
        in: body
        name: region
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateRegionDTO'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.RegionDTO'
      summary: Update region
      tags:
      - Locations
  /locations/search:
    get:
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.LocationSearchDTO'
      summary: Search across all location types
      tags:
      - Locations
  /locations/states:
    get:
      parameters:
      - description: Filter by country ID
        in: query
        name: country_id
        type: integer
      - description: Search by name
        in: query
        name: search
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.StateDTO'
            type: array
      summary: Get all states
      tags:
      - Locations
    post:
      parameters:
      - description: State data
        in: body
        name: state
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateStateDTO'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.StateDTO'
      summary: Create state
      tags:
      - Locations
  /locations/states/{id}:
    delete:
      parameters:
      - description: State ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: Delete state
      tags:
      - Locations
    get:
      parameters:
      - description: State ID
        in: path
        name: id
        required: true
        type: integer
      - description: Include cities
        in: query
        name: include_cities
        type: boolean
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.StateDTO'
      summary: Get state by ID
      tags:
      - Locations
    put:
      parameters:
      - description: State ID
        in: path
        name: id
        required: true
        type: integer
      - description: State data
        in: body
        name: state
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateStateDTO'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.StateDTO'
      summary: Update state
      tags:
      - Locations
  /locations/subregions:
    get:
      parameters:
      - description: Filter by region ID
        in: query
        name: region_id
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.SubregionDTO'
            type: array
      summary: Get all subregions
      tags:
      - Locations
    post:
      parameters:
      - description: Subregion data
        in: body
        name: subregion
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateSubregionDTO'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.SubregionDTO'
      summary: Create subregion
      tags:
      - Locations
  /locations/subregions/{id}:
    delete:
      parameters:
      - description: Subregion ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: Delete subregion
      tags:
      - Locations
    get:
      parameters:
      - description: Subregion ID
        in: path
        name: id
        required: true
        type: integer
      - description: Include countries
        in: query
        name: include_countries
        type: boolean
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.SubregionDTO'
      summary: Get subregion by ID
      tags:
      - Locations
    put:
      parameters:
      - description: Subregion ID
        in: path
        name: id
        required: true
        type: integer
      - description: Subregion data
        in: body
        name: subregion
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateSubregionDTO'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.SubregionDTO'
      summary: Update subregion
      tags:
      - Locations
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"